Am finalizat crud si pentru menu items

+++ src/main/java/com/eliasdejan/digial_menu/repository/MenuItemTypeRepository.java
@@ -3,5 +3,5 @@ package com.eliasdejan.digial_menu.repository;
 import com.eliasdejan.digial_menu.model.MenuItemType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuItemTypeRepository extends CrudRepository<MenuItemType, Long> {
+public interface MenuItemTypeRepository extends CrudRepository<MenuItemType, Integer> {
 }
